refactor: CardNickNameForm domain 디렉토리 이관
