+++ src/components/common/index.tsx
@@ -1,2 +1,1 @@
-export { default as CardNickNameForm } from './CardNickNameForm';
 export { default as TextInput } from './TextInput';
