+++ src/components/domain/index.ts
@@ -1,6 +1,7 @@
 export { default as CardCompanyModal } from './CardCompanyModal';
 export { default as CardCVCInput } from './CardCVCInput';
 export { default as CardForm } from './CardForm';
+export { default as CardNickNameForm } from './CardNickNameForm';
 export { default as CardNumberInput } from './CardNumberInput';
 export { default as CardOwnerInput } from './CardOwnerInput';
 export { default as CardPwdInput } from './CardPwdInput';
